webapp/requirements.txt
ipykernel==6.29
numpy==1.26
pandas==2.1
jupyter==1.0
tqdm==4.66
gunicorn==21.2
dash[celery]==2.18.2
dash-extensions==1.0.20
celery[redis]==5.4.0
dash-mantine-components==0.15
dash-iconify==0.1.2
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
dateparser==1.2.0
Wand==0.6.13
pyspellchecker==0.8.2
cloudscraper==1.2.71
selenium==4.27.1
